# docs/swagger.yaml
basePath: /
definitions:
  helper.BaseResponse:
    properties:
      data: {}
      errors: {}
      status:
        type: integer
    type: object
  helper.ExampleErrorResponse:
    properties:
      email:
        example: cannot be blank
        type: string
      message:
        example: user bad request
        type: string
      password:
        example: cannot be blank
        type: string
    type: object
...
  modelsocialmedia.Request:
    properties:
      name:
        type: string
      social_media_url:
        type: string
    type: object
  modelsocialmedia.Response:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      social_media_url:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  modelsocialmedia.ResponseList:
    properties:
      created_at:
        example: "2021-11-03T01:52:41.035Z"
        type: string
      id:
        example: 1
        type: integer
      name:
        example: jhondoe
        type: string
      social_media_url:
        example: https://example.com/url
        type: string
      updated_at:
        example: "2021-11-03T01:52:41.035Z"
        type: string
      user:
        properties:
          id:
            example: 1
            type: integer
          profile_image_url:
            example: https://example.com/photo.jpg
            type: string
          username:
            example: jhondoe
            type: string
        type: object
      user_id:
        example: 1
        type: integer
    type: object
  modelsocialmedia.ResponseListWrapper:
    properties:
      social_medias:
        items:
          $ref: '#/definitions/modelsocialmedia.ResponseList'
        type: array
    type: object
  modeluser.ExampleRequestUpdate:
    properties:
      email:
        example: test@example.com
        type: string
      username:
        example: jhondoe
        type: string
    type: object
  modeluser.ExampleResponseDelete:
    properties:
      message:
        example: your account has been successfully deleted
        type: string
    type: object
  modeluser.Request:
    properties:
      age:
        example: 23
        type: integer
      email:
        example: test@example.com
        type: string
      password:
        example: password
        type: string
      username:
        example: jhondoe
        type: string
    type: object
  modeluser.RequestLogin:
    properties:
      email:
        example: test@example.com
        type: string
      password:
        example: password
        type: string
    type: object
  modeluser.Response:
    properties:
      age:
        example: 23
        type: integer
      email:
        example: test@example.com
        type: string
      id:
        example: 1
        type: integer
      updated_at:
        example: "2021-11-03T01:52:41.035Z"
        type: string
      username:
        example: jhondoe
        type: string
    type: object
  modeluser.ResponseLogin:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJxd2Vxd2...
        type: string
    type: object
info:
  contact: {}
  description: This is API created, requirements for completing training Hacktiv8
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Hacktiv8 Final Project - MyGram API
  version: "1.0"
paths:
...
  /socialmedias:
    get:
      consumes:
      - application/json
      description: Get all social media
      parameters:
      - description: Bearer + user token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: SUCCESS
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/modelsocialmedia.ResponseListWrapper'
              type: object
      summary: Get all social media
      tags:
      - socialmedias
    post:
      consumes:
      - application/json
      description: Create social media
      parameters:
      - description: Bearer + user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/modelsocialmedia.Request'
      produces:
      - application/json
      responses:
        "201":
          description: CREATED
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/modelsocialmedia.Response'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
      summary: Create new social media
      tags:
      - socialmedias
  /socialmedias/:socialmediaid:
    delete:
      consumes:
      - application/json
      description: Delete by id social media
      parameters:
      - description: Bearer + user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID of the social media
        in: path
        name: socialmediaid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: SUCCESS
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/modelsocialmedia.Response'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
        "404":
          description: Record not found
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
      summary: Delete by id social media
      tags:
      - socialmedias
    put:
      consumes:
      - application/json
      description: Update by id social media
      parameters:
      - description: Bearer + user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID of the social media
        in: path
        name: socialmediaid
        required: true
        type: integer
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/modelsocialmedia.Request'
      produces:
      - application/json
      responses:
        "200":
          description: SUCCESS
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/modelsocialmedia.Response'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
        "404":
          description: Record not found
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
      summary: Update by id social media
      tags:
      - socialmedias
  /users:
    delete:
      consumes:
      - application/json
      description: Delete user
      parameters:
      - description: Bearer + user token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/modeluser.ExampleResponseDelete'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
        "401":
          description: Unauthorization
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
      summary: Delete user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user
      parameters:
      - description: Bearer + user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/modeluser.ExampleRequestUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/modeluser.Response'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
        "401":
          description: Unauthorization
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
      summary: Update user
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: Login user
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/modeluser.RequestLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/modeluser.ResponseLogin'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
        "404":
          description: Record not found
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
      summary: Login user
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: Create new user
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/modeluser.Request'
      produces:
      - application/json
      responses:
        "201":
          description: CREATED
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/modeluser.Response'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
        "409":
          description: data conflict, like email already exist
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
      summary: Create new user
      tags:
      - users
swagger: "2.0"